backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==4.0.2
attrs==22.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.5
Brotli==1.0.9
cachetools==5.2.1
certifi==2023.6.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==39.0.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
flatbuffers==2.0
fonttools==4.37.0
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.6.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
isoduration==20.11.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.8
libclang==16.0.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.5
mistune==3.1.3
mkl-service==2.4.0
mkl_fft==1.3.0
mkl_random==1.3.0
ml_dtypes==0.5.1
MouseInfo==0.1.3
multidict==6.0.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
opencv-contrib-python==4.11.0.86
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.0
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.7
plyer==2.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==5.9.3
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.11.3
pygetwindow==0.0.9
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyjwt==2.6.0
pyler==0.2.0
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytweening==1.2.0
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.21.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.5.0
uri-template==1.3.0
urllib3==1.26.11
uvicorn==0.21.1
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.3
win-inet-pton==1.1.0
win10toast==0.9
wrapt==1.14.1
yarl==1.8.2